Added max-lines to ESLint, created swagger-conf.js
